+++ src/helpers/nicetime.js
@@ -22,14 +22,14 @@ export default function nicetime(current, previous) {
         }
 
         else if (elapsed < msPerMonth) {
-            return 'approximately ' + Math.round(elapsed / msPerDay) + ' days ago';
+            return Math.round(elapsed / msPerDay) + ' days ago';
         }
 
         else if (elapsed < msPerYear) {
-            return 'approximately ' + Math.round(elapsed / msPerMonth) + ' months ago';
+            return Math.round(elapsed / msPerMonth) + ' months ago';
         }
 
         else {
-            return 'approximately ' + Math.round(elapsed / msPerYear) + ' years ago';
+            return Math.round(elapsed / msPerYear) + ' years ago';
         }
 }