+++ src/pages/about.js
@@ -97,7 +97,7 @@ export const query = graphql`
   query AboutQuery {
     projects: allMarkdownRemark(
       sort: {fields: [frontmatter___date], order: DESC}, 
-      limit: 3
+      limit: 4
       filter:{frontmatter:{section:{eq: "project"}}}
     ) {
       totalCount