+++ src/templates/blog-post.js
@@ -1,52 +1,91 @@
-import React from "react";
+import React, { Component } from "react";
 import Link from "gatsby-link";
 import Img from "gatsby-image";
 import nicetime from '../helpers/nicetime';
-import Disqus from 'disqus-react';
 
 import kebabCase from "lodash/kebabCase";
 import 'prismjs/themes/prism-okaidia.css';
 
+import ReadingProgress from '../components/ReadingProgress';
 import SEO from '../components/SEO';
+import Cover from '../components/Cover';
+import Comments from '../components/Comments';
 import PostLoop from '../components/PostLoop';
 import Twitter from '../components/icons/Twitter'
 import Tumblr from '../components/icons/Tumblr'
 import ryosukeAvatar from '../assets/img/ryosuke-avatar-128.png'
 
-export default ({ data }) => {
+export default class BlogPost extends Component {
+
+  constructor(props) {
+    super(props);
+    
+    this.state = {
+      'instagram': false
+    };
+  }
+
+  componentDidMount() {
+    if (window.instgrm || document.getElementById('react-instagram-embed-script')) {
+      if(this.state.instagram == false)
+      {
+        window.instgrm.Embeds.process()
+      }
+    } else {
+
+      // Create script element with Instagram embed JS lib
+      const s = document.createElement('script')
+      s.async = s.defer = true
+      s.src = `//www.instagram.com/embed.js`
+      s.id = 'react-instagram-embed-script'
+      const body: HTMLElement | null = document.body
+      if (body) {
+        body.appendChild(s)
+      }
+
+      // Run Instagram function to show embeds
+      if (window.instgrm && this.state.instagram == false) {
+        window.instgrm.Embeds.process()
+      }
+
+      // Set IG state to true so the process doesn't run again
+      this.setState({
+        'instagram': true
+      });
+    }
+  }
+
+  render() {
     const skip = false;
-    const post = data.blog;
+    const post = this.props.data.blog;
     let related;
-    data.relatedPosts ? related = data.relatedPosts.edges : related = null;
+    this.props.data.relatedPosts ? related = this.props.data.relatedPosts.edges : related = null;
     const currentDate = new Date();
 
     const tags = post.frontmatter.tags.map((tag) => (
       <li key={tag}><Link to={'/tags/' + kebabCase(tag) }>#{ tag }</Link></li>
     ));
 
-    const postImage = post.frontmatter.cover_image.childImageSharp && post.frontmatter.cover_image.childImageSharp.sizes && post.frontmatter.cover_image.childImageSharp.sizes.src;
+    let postImage = post.frontmatter.cover_image.publicURL;
+    let postDate = new Date(post.frontmatter.date);
 
-
-    const disqusShortname = 'ryosuke';
-    const disqusConfig = {
-      url: `http://ryosuke.design/${post.fields.slug}`,
-      identifier: post.fields.slug,
-      title: post.frontmatter.title,
-    };
+    if (post.frontmatter.cover_image.childImageSharp !== null) {
+      postImage = post.frontmatter.cover_image.childImageSharp && post.frontmatter.cover_image.childImageSharp.sizes && post.frontmatter.cover_image.childImageSharp.sizes.src;
+    }
 
     return (
         <div className="Blog">
+          {/*----- Reading progress only on blog -----*/}
+          { post.frontmatter.section === 'blog' && <ReadingProgress targetEl="#Article" /> }
           <SEO 
             key={`seo-${post.fields.slug}`}
             postImage={postImage}
             postData={post}
             isBlogPost
           />
-          <article className="ArticlePage">
-            {/*----- Cover image -----*/}
-            <figure className="Cover">
-              <Img sizes={post.frontmatter.cover_image.childImageSharp.sizes} />
-            </figure>
+          <article className={"ArticlePage " + post.frontmatter.section } id="Article">
+            {/*----- Cover image only on blog -----*/}
+            { post.frontmatter.section === 'blog' && <Cover image={post.frontmatter.cover_image} /> }
             <section className="container">
 
               {/*----- Post content -----*/}
@@ -69,16 +108,16 @@ export default ({ data }) => {
                   <img src={ryosukeAvatar} alt="Blue square avatar white centered hiragana text reading Ryosuke" />
                   <h5>
                     @Ryosuke
-                    <span className="date">{nicetime(currentDate, new Date(post.frontmatter.date))}</span>                  
+                    <span className="date">{ nicetime(currentDate, postDate) }</span>                  
                   </h5>
                 </figure>
                 <section className="share">
-                  <a href={`http://twitter.com/share?text=${post.frontmatter.title}&url=http://whoisryosuke.github.io/ryosyke-gatsby-blog/${post.fields.slug}&hashtags=${post.frontmatter.tags }`} className="twitter">
-                    Share on Twitter
+                  <a href={`http://twitter.com/share?text=${post.frontmatter.title}&url=http://whoisryosuke.com/${post.fields.slug}&hashtags=${post.frontmatter.tags }`} className="twitter">
+                    { post.frontmatter.section === 'blog' ? 'Share on Twitter' : 'Share' }
                     <Twitter />
                   </a>
-                  <a href={`http://www.tumblr.com/share/link?url=http://whoisryosuke.github.io/ryosyke-gatsby-blog${post.fields.slug}`} className="tumblr">
-                    Share on Tumblr
+                  <a href={`http://www.tumblr.com/share/link?url=http://whoisryosuke.com${post.fields.slug}`} className="tumblr">
+                    { post.frontmatter.section === 'blog' ? 'Share on Tumblr' : 'Share' }
                     <Tumblr />
                   </a>
                 </section>
@@ -88,12 +127,7 @@ export default ({ data }) => {
 
         </article>
 
-        <section className="Comments container">
-          <section className="Segment content">
-            <h3 className="Title">Leave a comment</h3>            
-            <Disqus.DiscussionEmbed shortname={disqusShortname} config={disqusConfig} />
-          </section>
-        </section>
+        { post.frontmatter.section === 'blog' && <Comments post={post} /> }
 
           { related ? 
             <nav className="RelatedPosts container">
@@ -105,6 +139,7 @@ export default ({ data }) => {
           }
         </div>
     );
+  }
 };
 
 export const query = graphql`
@@ -114,6 +149,7 @@ export const query = graphql`
       frontmatter {
         title
         cover_image {
+              publicURL
               childImageSharp {
                 sizes(maxWidth: 1240 ) {
                   tracedSVG
@@ -124,6 +160,7 @@ export const query = graphql`
             }
         date(formatString: "DD MMMM, YYYY")
         tags
+        section
       }
       fields {
         slug
@@ -143,6 +180,7 @@ export const query = graphql`
           frontmatter {
             title
             cover_image {
+                  publicURL
                   childImageSharp {
                     sizes(maxWidth: 1240 ) {
                       src